check auth with cookie & set cookie on login
